Exploring how to handle multiple outputs

+++ examples/two_eq_lag.rs
@@ -35,7 +35,7 @@ impl ode_solvers::System<State> for Model<'_> {
 
         //////////////// END USER DEFINED ////////////////
         for dose in &self.scenario.doses{
-            if (t-dose.time as f64).abs() < 1.0e-4 {
+            if (t-dose.time).abs() < 1.0e-4 {
                 y[dose.compartment] += dose.dose;
             }
         }
